API/requirements.txt
absl-py==2.1.0
anaconda==0.0.1.1
annotated-types==0.6.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
easyocr==1.7.1
fastapi==0.110.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
h11==0.14.0
h5py==3.11.0
idna==3.6
imageio==2.34.0
jsonify==0.5
keras==3.2.1
kiwisolver==1.4.5
labelbox==3.63.0
labelImg==1.8.6
lazy_loader==0.4
libclang==18.1.1
libs==0.0.10
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-image==0.23.1
scipy==1.13.0
seaborn==0.13.2
setuptools==69.2.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.2.12
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.47
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0